+++ src/registry/index.test.ts
@@ -22,54 +22,7 @@ describe('Register module', () => {
 		},
 	});
 
-	const routes = controllerRegistry()
-		.register(UserController, {
-			index: async () => {
-				return [
-					{
-						age: 1,
-						name: 'a',
-					},
-				];
-			},
-			read: async () => {
-				return {
-					age: 2,
-					name: 'b',
-				};
-			},
-			create: async () => {
-				return {
-					age: 3,
-					name: 'c',
-				};
-			},
-			update: async () => {
-				return {
-					age: 4,
-					name: 'd',
-				};
-			},
-			destroy: async () => {
-				return {
-					age: 5,
-					name: 'e',
-				};
-			},
-			statistics: {
-				index: async () => {
-					return {
-						stat: 10,
-					};
-				},
-				read: async (data: CRUDBase) => {
-					return {
-						stat: data.id as number,
-					};
-				},
-			},
-		})
-		.handle();
+	const routes = controllerRegistry().register(UserController, UserController).handle();
 
 	test('Register handle', () => {
 		expect(routes.has('users')).toBeTruthy();